Style Notifications View
