+++ UCLA Radio/UCLA Radio/extensions/DateFormatter+ShowDate.swift
@@ -10,14 +10,16 @@ import Foundation
 
 extension DateFormatter {
     
+    // See https://developer.apple.com/library/content/documentation/Cocoa/Conceptual/DataFormatting/Articles/dfDateFormatting10_4.html for date formatting
     enum DateFormat: String {
-        case DayAndTime = "EEE ha"
-        case Time = "ha"
+        case DayAndHour = "EEE ha"
+        case Hour = "ha"
+        case HourAndMinute = "h:mma"
     }
     
     // Expect string of DayAndTime DateFormat ("EEE ha")
     func formatShowTimeStringToDateComponents(_ s: String) -> DateComponents? {
-        dateFormat = DateFormat.DayAndTime.rawValue
+        dateFormat = DateFormat.DayAndHour.rawValue
         if let date = self.date(from: s) {
             var components = Calendar(identifier: .gregorian).dateComponents([.hour, .weekday], from: date)
             components.timeZone = TimeZone(identifier: "America/Los_Angeles")
